feat: 삭제하려는 일정 정보가 이미 삭제 상태일 때 예외
DeletedTodoException 발생

+++ src/main/java/com/sparta/springbasicsolo/service/TodoService.java
@@ -56,6 +56,9 @@ public class TodoService {
     public void deleteTodo(Long id, String password) {
         Optional<TodoDTO> findTodo = todoRepository.findById(id);
         if (findTodo.isPresent()) {
+            if (findTodo.get().getIsDeleted()) {
+                throw new DeletedTodoException("이미 삭제된 일정입니다.");
+            }
             if (Objects.equals(findTodo.get().getPassword(), password)) {
                 log.info("비밀번호 일치함. 삭제 진행");
                 todoRepository.deleteTodo(id);
